requirements.txt
functions-framework==3.*
pandas
gspread
google-cloud-storage
google-cloud-bigquery
requests
python-dotenv
